[REFACTOR] move file swaggerExceptionResponsesAdaptor annotation -> adaptor
