+++ src/main/kotlin/com/tinuproject/tinu/swagger/adaptor/SwaggerExceptionResponseAdaptor.kt
@@ -1,8 +1,9 @@
-package com.tinuproject.tinu.swagger.annotation
+package com.tinuproject.tinu.swagger.adaptor
 
 import com.tinuproject.tinu.DTO.ErrorResponse
 import com.tinuproject.tinu.DTO.ResponseDTO
 import com.tinuproject.tinu.domain.exception.base.BaseException
+import com.tinuproject.tinu.swagger.annotation.ExampleHolder
 import io.swagger.v3.oas.models.Operation
 import io.swagger.v3.oas.models.examples.Example
 import io.swagger.v3.oas.models.media.Content
